+++ core/src/consensus/mod.rs
@@ -890,8 +890,14 @@ impl ConsensusGraph {
     ) -> RpcResult<ExecutionOutcome> {
         // only allow to call against stated epoch
         self.validate_stated_epoch(&epoch)?;
-        let epoch_id = self.get_hash_from_epoch_number(epoch)?;
-        self.executor.call_virtual(tx, &epoch_id)
+        let (epoch_id, epoch_size) = if let Ok(v) =
+            self.get_block_hashes_by_epoch(epoch)
+        {
+            (v.last().expect("pivot block always exist").clone(), v.len())
+        } else {
+            bail!("cannot get block hashes in the specified epoch, maybe it does not exist?");
+        };
+        self.executor.call_virtual(tx, &epoch_id, epoch_size)
     }
 
     pub fn check_balance_against_transaction(